consolidated static data
