+++ src/components/aboutMeComponents/defaultDisplays/contactInfoDisplay.js
@@ -17,7 +17,7 @@ import {
   QuestionWithMultipleAnswers,
 } from './QuestionAnswerDisplayTemplates.js';
 import {getListOfAnswers} from 'lib/helperFunctions/helpers';
-import {raceLabels, blankProfile} from '../others/defaultData';
+import {raceLabels, blankProfile} from '../../../lib/formHelpers/defaultData';
 
 const ContactInfoDisplay = ({contact, isOnEditMode, onClickEdit, classes}) => {
   const {first_name, last_name, phone_primary, email} = contact;
