+++ src/components/aboutMeComponents/forms/ProgramsAndEligibilityForm.container.js
@@ -6,7 +6,7 @@ import {
 } from 'state/contacts/contacts.actions';
 import {getAllProgramNames} from 'state/programs';
 import ProgramsAndEligibilityForm from './ProgramsAndEligibilityForm';
-import {defaultPrograms} from '../others/defaultData';
+import {defaultPrograms} from '../../../lib/formHelpers/defaultData';
 
 export const mapStateToProps = state => {
   const programs = Object.values(state.programs);
