+++ src/lib/formHelpers/defaultData.js
@@ -450,6 +450,29 @@ const defaultPrograms = [
   },
 ];
 
+const monthFullNames = [
+  'January',
+  'February',
+  'March',
+  'April',
+  'May',
+  'June',
+  'July',
+  'August',
+  'September',
+  'October',
+  'November',
+  'December',
+];
+
+// Creating an array of current year to 1900
+let currentYear = new Date();
+currentYear = currentYear.getFullYear();
+let years = [];
+for (let i = currentYear; i > 1900; i--) {
+  years.push(i);
+}
+
 export {
   countryList,
   states,
@@ -466,4 +489,6 @@ export {
   hearAboutUsOptions,
   blankProfile,
   defaultPrograms,
+  monthFullNames,
+  years
 };
